gator/pyproject.toml
[project]
name = "gator"
description = "Simple Component-Driven Game Engine"
version = "0.0.1"
readme = "README.md"

authors = [
    { name="Damus666", email="macchia.cuordileone@gmail.com" }
]

dependencies = [
    "glfw>=2.6.2",
    "imgui>=2.0.0",
    "numpy>=1.24.3",
    "Pillow>=10.0.0",
    "PyOpenGL>=3.1.6",
    "PyGLM>=2.7.0"
]

[project.scripts]

[project.urls]
homepage = "https://github.com/Damus666/GatorEngine"
bugtracker = "https://github.com/Damus666/GatorEngine/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-rA -s"
testpaths = []